// src/app/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  EMPLOYEE: 'EMPLOYEE',
  ADMIN: 'ADMIN',
  TECHNICIAN: 'TECHNICIAN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const MaintenanceFor = {
  EQUIPMENT: 'EQUIPMENT',
  WORK_CENTER: 'WORK_CENTER'
} as const

export type MaintenanceFor = (typeof MaintenanceFor)[keyof typeof MaintenanceFor]


export const MaintenanceType = {
  CORRECTIVE: 'CORRECTIVE',
  PREVENTIVE: 'PREVENTIVE'
} as const

export type MaintenanceType = (typeof MaintenanceType)[keyof typeof MaintenanceType]


export const RequestStatus = {
  NEW: 'NEW',
  ASSIGNED: 'ASSIGNED',
  IN_PROGRESS: 'IN_PROGRESS',
  REPAIRED: 'REPAIRED',
  SCRAPPED: 'SCRAPPED'
} as const

export type RequestStatus = (typeof RequestStatus)[keyof typeof RequestStatus]


export const Priority = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
  CRITICAL: 'CRITICAL'
} as const

export type Priority = (typeof Priority)[keyof typeof Priority]